% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_figs.R
\name{cnetfilt}
\alias{cnetfilt}
\title{Filter a gene set enrichment result to eliminate genes with low absolute fold-change}
\usage{
cnetfilt(gse, fcut = 2, showcat = 3)
}
\arguments{
\item{gse}{gene-set enrichment object from clusterprofiler}

\item{fcut}{minimum absolute foldchange of genes to be retained}

\item{showcat}{number of gene pathways to plot}
}
\value{
cnetplot with filtered genes
}
\description{
Filter a gene set enrichment result to eliminate genes with low absolute fold-change
}
\examples{
todo
}
