% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetonic_msig.R
\name{limma2df}
\alias{limma2df}
\title{convert limma results to dataframe with DESeq column names}
\usage{
limma2df(res_de, FDR = NULL)
}
\arguments{
\item{res_de}{limma toptable results}

\item{FDR}{false discovery rate cutoff}
}
\value{
dataframe with deseq-like result column names
}
\description{
convert limma results to dataframe with DESeq column names
}
\examples{
TODO
}
