% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_figs.R
\name{rnaplots}
\alias{rnaplots}
\title{Title}
\usage{
rnaplots(dds, folder = NULL, fprefix = NULL)
}
\arguments{
\item{dds}{DESeqDataset object}

\item{folder}{output result folder}

\item{fprefix}{output analysis prefix}
}
\value{
list of plots
}
\description{
Title
}
\examples{
todo
}
