% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetonic_msig.R
\name{shake_gsvaResult}
\alias{shake_gsvaResult}
\title{prepare gsva results (after limma DE) for downstream genetonic}
\usage{
shake_gsvaResult(obj, res_de, m_df, gset, anno_df)
}
\arguments{
\item{obj}{limma toptable with gsva as input}

\item{res_de}{DE dataframe with ensembl ids}

\item{m_df}{annotations from msigdbr}

\item{gset}{genesets from gsva results (geneSets(gsva_res))}

\item{anno_df}{dataframe with ensembl to symbol mapping}
}
\value{
dataframe with standardized names for downstream genetonic
}
\description{
prepare gsva results (after limma DE) for downstream genetonic
}
\examples{
TODO
}
