% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_figs.R
\name{rnaplots}
\alias{rnaplots}
\title{automagically make a bunch of standard deseq2 and clusterprofiler plots}
\usage{
rnaplots(dds, pcut = 0.05, folder = NULL, fprefix = NULL)
}
\arguments{
\item{dds}{DESeqDataset object}

\item{pcut}{gsea pvalue threshold}

\item{folder}{output result folder}

\item{fprefix}{output analysis prefix}
}
\value{
list of plots
}
\description{
automagically make a bunch of standard deseq2 and clusterprofiler plots
}
\examples{
todo
}
